requirements.txt
# Qwen Image Service - Dependencies
# ==================================

# Web框架
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# 配置管理
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyYAML>=6.0.1

# AI/ML
# torch 已包含在基础镜像中，无需重复安装
# torch>=2.4.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
sentencepiece>=0.2.0

# 图像处理
Pillow>=10.0.0

# 可选: 环境变量管理
python-dotenv>=1.0.0
